Major updates on elastic net and non-compositional covariates (#11)
* May15-elnet+covariates

* Optimize two-step procedure for non-compositional covariates

* bugfix for 2-step selection

* Adding multiple k-fold cross-validation option.

* Adding function to compare `a`.

* Adding new dependencies

* Allow penalty-weighting for non-compositional features.

* Optimizing step2 output / parallel computing / loss function for elastic net with non-compositional covariates.

* minor bug fix for logistic regression a.FLORAL() and two-step output when there's no compositional features selected.

* Update version number for PR

* bugfix for scenarios where only one taxa is selected (which causes invalid results because there must be a pair of taxa to form a ratio).

* mcv.FLORAL: output mean coefficients out of multiple cv runs.

* fix issues with empty 2-stage tables.

* New plot options for mcv.FLORAL. Minor improvements for ridge regression.
