+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: FLORAL
 Type: Package
 Title: Fit Log-Ratio Lasso Regression for Compositional Data
-Version: 0.1.0.9000
+Version: 0.1.0.9001
 Date: 2023-05-09
 Authors@R: c(
     person("Teng", "Fei", , "feit1@mskcc.org", role = c("aut", "cre", "cph"),
@@ -32,7 +32,10 @@ Imports:
   caret,
   grDevices,
   utils,
-  mvtnorm
+  mvtnorm,
+  doParallel,
+  doRNG,
+  foreach
 LinkingTo: Rcpp, RcppArmadillo, RcppProgress
 RoxygenNote: 7.2.3
 Encoding: UTF-8
