+++ NAMESPACE
@@ -1,11 +1,17 @@
 # Generated by roxygen2: do not edit by hand
 
 export(FLORAL)
+export(a.FLORAL)
+export(mcv.FLORAL)
 export(simu)
 import(Rcpp)
+import(doParallel)
+import(doRNG)
 import(dplyr)
+import(foreach)
 import(ggplot2)
 import(glmnet)
+import(parallel)
 import(survival)
 importFrom(caret,createFolds)
 importFrom(grDevices,rainbow)
