+++ inst/WORDLIST
@@ -1,12 +1,12 @@
 Biostatistics
 CMD
 CRC
-Codecov
 Compositional
 DM
 Dai
 Devlin
 Gönen
+Interpretable
 JU
 LOg
 Lv
@@ -28,8 +28,11 @@ et
 feit
 finegray
 https
+microbiome
 mskcc
 multivariable
+ncv
 preprint
+reproducibility
 se
 λ
