+++ man/a.FLORAL.Rd
@@ -1,10 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/FLORAL.R
-\name{FLORAL}
-\alias{FLORAL}
-\title{Fit Log-ratio lasso regression for compositional covariates}
+\name{a.FLORAL}
+\alias{a.FLORAL}
+\title{Comparing prediction performances under different choices of weights for lasso/ridge penalty}
 \usage{
-FLORAL(
+a.FLORAL(
+  a = c(0.1, 0.5, 1),
+  ncore = 1,
+  seed = NULL,
   x,
   y,
   ncov = 0,
@@ -16,17 +19,20 @@ FLORAL(
   length.lambda = 100,
   lambda.min.ratio = NULL,
   ncov.lambda.weight = 0,
-  a = 1,
   mu = 1,
   ncv = 5,
   intercept = FALSE,
-  foldid = NULL,
-  step2 = TRUE,
-  progress = TRUE,
-  plot = TRUE
+  step2 = FALSE,
+  progress = TRUE
 )
 }
 \arguments{
+\item{a}{vector of scalars between 0 and 1 for comparison.}
+
+\item{ncore}{Number of cores used for parallel computation. Default is to use only 1 core.}
+
+\item{seed}{A random seed for reproducibility of the results. By default the seed is the numeric form of \code{Sys.Date()}.}
+
 \item{x}{Feature matrix, where rows specify subjects and columns specify features. The first \code{ncov} columns should be patient characteristics and the rest columns are microbiome absolute counts corresponding to various taxa. If \code{x} contains longitudinal data, the rows must be sorted in the same order of the subject IDs used in \code{y}.}
 
 \item{y}{Outcome. For a continuous or binary outcome, \code{y} is a vector. For survival outcome, \code{y} is a \code{Surv} object.}
@@ -49,48 +55,28 @@ FLORAL(
 
 \item{ncov.lambda.weight}{Weight of the penalty lambda applied to the first \code{ncov} covariates. Default is 0 such that the first \code{ncov} covariates are not penalized.}
 
-\item{a}{A scalar between 0 and 1: \code{a} is the weight for lasso penalty while \code{1-a} is the weight for ridge penalty.}
-
 \item{mu}{Value of penalty for the augmented Lagrangian}
 
 \item{ncv}{Folds of cross-validation. Use \code{NULL} if cross-validation is not wanted.}
 
 \item{intercept}{\code{TRUE} or \code{FALSE}, indicating whether an intercept should be estimated.}
 
-\item{foldid}{A vector of fold indicator. Default is \code{NULL}.}
-
 \item{step2}{\code{TRUE} or \code{FALSE}, indicating whether a second-stage feature selection for specific ratios should be performed for the features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}
 
 \item{progress}{\code{TRUE} or \code{FALSE}, indicating whether printing progress bar as the algorithm runs.}
-
-\item{plot}{\code{TRUE} or \code{FALSE}, indicating whether returning plots of model fitting.}
 }
 \value{
-A list with path-specific estimates (beta), path (lambda), and others. Details can be found in \code{README.md}.
+A \code{ggplot2} object of cross-validated prediction metric versus \code{lambda}, stratified by \code{a}. Detailed data can be retrieved from the \code{ggplot2} object itself.
 }
 \description{
-Conduct log-ratio lasso regression for continuous, binary and survival outcomes.
+Summarizing \code{FLORAL} outputs from various choices of \code{a}
 }
 \examples{
 
 set.seed(23420)
 
-# Continuous outcome
 dat <- simu(n=50,p=30,model="linear")
-fit <- FLORAL(dat$xcount,dat$y,family="gaussian",progress=FALSE,step2=TRUE)
-
-# Binary outcome
-dat <- simu(n=50,p=30,model="binomial")
-fit <- FLORAL(dat$xcount,dat$y,family="binomial",progress=FALSE,step2=TRUE)
-
-# Survival outcome
-dat <- simu(n=50,p=30,model="cox")
-fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d),family="cox",progress=FALSE,step2=TRUE)
-
-# Competing risks outcome
-dat <- simu(n=50,p=30,model="finegray")
-fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d,type="mstate"),failcode=1,
-              family="finegray",progress=FALSE,step2=FALSE)
+pmetric <- a.FLORAL(a=c(0.1,1),ncore=1,x=dat$xcount,y=dat$y,family="gaussian",progress=FALSE)
 
 }
 \references{
