+++ man/mcv.FLORAL.Rd
@@ -1,10 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/FLORAL.R
-\name{FLORAL}
-\alias{FLORAL}
-\title{Fit Log-ratio lasso regression for compositional covariates}
+\name{mcv.FLORAL}
+\alias{mcv.FLORAL}
+\title{Summarizing selected compositional features over multiple cross validations}
 \usage{
-FLORAL(
+mcv.FLORAL(
+  mcv = 10,
+  ncore = 1,
+  seed = NULL,
   x,
   y,
   ncov = 0,
@@ -20,13 +23,18 @@ FLORAL(
   mu = 1,
   ncv = 5,
   intercept = FALSE,
-  foldid = NULL,
   step2 = TRUE,
   progress = TRUE,
   plot = TRUE
 )
 }
 \arguments{
+\item{mcv}{Number of random `ncv`-fold cross-validation to be performed.}
+
+\item{ncore}{Number of cores used for parallel computation. Default is to use only 1 core.}
+
+\item{seed}{A random seed for reproducibility of the results. By default the seed is the numeric form of \code{Sys.Date()}.}
+
 \item{x}{Feature matrix, where rows specify subjects and columns specify features. The first \code{ncov} columns should be patient characteristics and the rest columns are microbiome absolute counts corresponding to various taxa. If \code{x} contains longitudinal data, the rows must be sorted in the same order of the subject IDs used in \code{y}.}
 
 \item{y}{Outcome. For a continuous or binary outcome, \code{y} is a vector. For survival outcome, \code{y} is a \code{Surv} object.}
@@ -57,40 +65,24 @@ FLORAL(
 
 \item{intercept}{\code{TRUE} or \code{FALSE}, indicating whether an intercept should be estimated.}
 
-\item{foldid}{A vector of fold indicator. Default is \code{NULL}.}
-
 \item{step2}{\code{TRUE} or \code{FALSE}, indicating whether a second-stage feature selection for specific ratios should be performed for the features selected by the main lasso algorithm. Will only be performed if cross validation is enabled.}
 
 \item{progress}{\code{TRUE} or \code{FALSE}, indicating whether printing progress bar as the algorithm runs.}
 
-\item{plot}{\code{TRUE} or \code{FALSE}, indicating whether returning plots of model fitting.}
+\item{plot}{\code{TRUE} or \code{FALSE}, indicating whether returning summary plots of selection probability for taxa features.}
 }
 \value{
-A list with path-specific estimates (beta), path (lambda), and others. Details can be found in \code{README.md}.
+A list with relative frequencies of a certain feature being selected over \code{mcv} \code{ncv}-fold cross-validations.
 }
 \description{
-Conduct log-ratio lasso regression for continuous, binary and survival outcomes.
+Summarizing \code{FLORAL} outputs from multiple random k-fold cross validations
 }
 \examples{
 
 set.seed(23420)
 
-# Continuous outcome
 dat <- simu(n=50,p=30,model="linear")
-fit <- FLORAL(dat$xcount,dat$y,family="gaussian",progress=FALSE,step2=TRUE)
-
-# Binary outcome
-dat <- simu(n=50,p=30,model="binomial")
-fit <- FLORAL(dat$xcount,dat$y,family="binomial",progress=FALSE,step2=TRUE)
-
-# Survival outcome
-dat <- simu(n=50,p=30,model="cox")
-fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d),family="cox",progress=FALSE,step2=TRUE)
-
-# Competing risks outcome
-dat <- simu(n=50,p=30,model="finegray")
-fit <- FLORAL(dat$xcount,survival::Surv(dat$t,dat$d,type="mstate"),failcode=1,
-              family="finegray",progress=FALSE,step2=FALSE)
+fit <- mcv.FLORAL(mcv=2,ncore=1,x=dat$xcount,y=dat$y,family="gaussian",progress=FALSE,step2=TRUE)
 
 }
 \references{
